Receiving messages over UART

+++ STM32F7_Blinky/app/inc/comm.h
@@ -33,7 +33,7 @@
 void    COMM_Init(int baud);
 void    COMM_Putc(char c);
 char    COMM_Getc(void);
-uint8_t COMM_GetFrame(uint8_t* buf, uint8_t* len);
+int     COMM_GetFrame(uint8_t* buf, uint8_t* len);
 
 /**
  * @}
